Renamed rate_limiting with prefix utils_ to distinguish from app module
